requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
python-telegram-bot==20.7
mercadopago==2.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.3
PyPDF2==3.0.1
openpyxl==3.1.2
google-api-python-client==2.109.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
schedule==1.2.0
openai==1.3.7
anthropic==0.8.1
pgvector==0.2.4
sentence-transformers==2.2.2
numpy==1.24.3
#pip install psycopg2-binary python-dotenv
#pip install --upgrade sentence-transformers>=2.3.0 diffusers>=0.29.0 huggingface_hub>=0.26.0
